+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.0.0 (unreleased)
 ------------------
 
+- #1778 Added Datamanager Adapters for Analysis and Sample
 - #1777 Allow to re-add cancelled/rejected/retracted analyses to a sample
 - #1777 Fix APIError when a retest analysis source was removed from a sample
 - #1776 Move time calculation into cache key function